core: rework page-fault resolution
The change "core: allow offset-attached managed dataspaces" addressed a
corner case of the use of nested region maps. Apparently, this change
negatively affects other scenarios (tool_chain_auto).

In order to confidently cover all the differnt situations, this patch
reworks the page-fault resolution code for improved clarity and safety,
by introducing dedicated result types, reducing the use of basic types,
choosing expressive names, and fostering constness.

It also introduces a number of 'print' hooks that greatly ease manual
instrumentation and streamlines the error messages printed by core.
Those messages no longer appear when a user-level page-fault handler
is reistered for the faulted-at region map. So the monitor component
produces less noise on the attempt to dump non-existing memory.

Issue #4917
Fixes #4920

+++ repos/base/src/include/base/internal/page_size.h
@@ -18,8 +18,8 @@
 
 namespace Genode
 {
-	constexpr size_t get_page_size_log2() { return 12; }
-	constexpr size_t get_page_size()      { return 1 << get_page_size_log2(); }
+	constexpr uint8_t get_page_size_log2() { return 12; }
+	constexpr size_t  get_page_size()      { return 1 << get_page_size_log2(); }
 }
 
 #endif /* _INCLUDE__BASE__INTERNAL__PAGE_SIZE_H_ */
